+++ docs/changes.md
@@ -25,6 +25,7 @@
 - Manifest v3 has much better error handling and there should be less uncaptured errors.
 - Fixed font size adjustment for 2x and 3x values. Text now centre aligns and stopped unexpected text wrapping.
 - Fixed Options reloading the page when a new tab was selected.
+- New Toolbar icon setting to select dark or light mode button.
 
 ## 4.2